backend/fastapi-app/requirements.txt
# Framework principal
fastapi==0.104.1
uvicorn==0.24.0

# Machine Learning
pandas==2.0.3
scikit-learn==1.3.2
joblib==1.3.2

# Validaciones y multipart
pydantic==2.5.0
python-multipart==0.0.6

# Scheduler para reentrenamiento
schedule==1.2.0

# Seguridad y autenticación
bcrypt==4.0.1
PyJWT==2.8.0

# Base de datos
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
motor==3.3.2   # opcional: MongoDB para logs/eventos

# Testing
pytest==7.4.3
httpx==0.25.2
